+++ cli/package.json
@@ -32,7 +32,6 @@
         "@types/babel__generator": "^7.6.8",
         "@types/babel__traverse": "^7.20.6",
         "@types/bun": "latest",
-        "@types/degit": "^2.8.6",
         "esbuild": "^0.23.1",
         "tslib": "^2.6.3",
         "typescript": "^5.0.0"
@@ -43,7 +42,6 @@
         "@babel/traverse": "^7.14.5",
         "@babel/types": "^7.14.5",
         "commander": "^12.1.0",
-        "degit": "^2.8.4",
         "glob": "^11.0.0",
         "ora": "^8.1.0"
     }
