--- cli/src/create/constant.ts
@@ -1,1 +1,0 @@
-export const NEXT_TEMPLATE_REPO = 'onlook-dev/starter';
