+++ utils/src/models.ts
@@ -12,3 +12,12 @@ export type ErrorResponse = {
         code: string;
     };
 };
+
+export enum ProjectCreationStage {
+    CLONING = 'cloning',
+    INSTALLING = 'installing',
+    COMPLETE = 'complete',
+    ERROR = 'error'
+}
+
+export type ProgressCallback = (stage: ProjectCreationStage, message: string) => void;
